refactor server auth rate limit helpers
